requirements.txt
PyPDF2==3.0.1
google-generativeai==0.6.0
tqdm==4.65.0
langchain 
langchain-google-genai